Fix a Python 3 breakage
In Python 3, generators don't have a next() method; it's __next__().
The simple work-around is to use six.next(), but this requires an
extensive tweak to the test suite, because mock objects cannot
pretend to be generators.  (Well, they can, but without the level of
control needed to properly test cli_tools.)  This change thus also
adds MockGenFunc and its associated MockGen class, to allow proper
generator mocks.